requirements.txt
blinker==1.9.0
click==8.1.8
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.11.0
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
psycopg2==2.9.10
Pygments==2.19.1
python-dotenv==1.1.0
rich==13.9.4
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
